requirements.txt
setuptools
git-python
